Accomodate changes to authorization using SSO
Instead of relying on long-lived credentials being stored locally,
I now use aws-cli with sso. This commit modifies the "profile" of
each aws provider block so that is uses the "org" profile to make
changes. For changes in child accounts, we then assume the AWS-
managed OrganizationAccountAccessRole role.

This ensures that the governance/admin changes will not be blocked
by any userland changes (that role can't be changed or removed by
a non-root user). This gives me centralized control over all sub-
accounts from the organization root, and improves the security
posture of the organization as I can now enforce MFA and account
access rules/permissions from a single source.